Fix build issues on 32-bit platforms.
The string ID's are stored in mdds's unsigned long block, but we
have changed the underlying type of string ID's from unsigned long
to uint64_t.  Because of this, on 32-bit systems it caused a type
mismatch.  On 64-bit systems this was fine since unsigned long was
the underlying type for uint64_t...