Replace all uses of SPDLOG_DEBUG and _TRACE with IXION_*.
While spdlog is great, my need for logging is limited to compile
time and wasn't worth the extra hussle of having another run-time
dependency.  IXION's own logging is very simple, and is inspired by
the logging facility implemented in LibreOffice which was done by
Stephan Bergmann.  That said, the one in ixion is very simplified
version of LibreOffice's logging.